PR changes to sankey functions
