+++ usaon_vta_survey/util/full_sankey.py
@@ -1,10 +1,8 @@
-from itertools import chain
-
-from usaon_vta_survey.models.tables import (
-    Response,
-    ResponseApplication,
-    ResponseDataProduct,
-    ResponseObservingSystem,
+from usaon_vta_survey.models.tables import Response
+from usaon_vta_survey.util.sankey import (
+    _applications_sankey,
+    _data_products_sankey,
+    _societal_benefit_areas_sankey,
 )
 
 
@@ -17,81 +15,9 @@ def sankey(response: Response) -> list[list[object]]:
 
 
 def _sankey(response: Response) -> list[tuple[str, str, int]]:
-    """Provide a sankey data structure of applications, formatted for type checker."""
+    """Provide a sankey data structure of applications."""
     data1 = _applications_sankey(response)
     data2 = _data_products_sankey(response)
     data3 = _societal_benefit_areas_sankey(response)
     data = data1 + data2 + data3
-    return data
-
-
-def _societal_benefit_areas_sankey(response: Response):
-    data = list(
-        chain(
-            *[
-                _application_societal_benefit_area_sankey_links(application)
-                for application in response.applications
-            ]
-        )
-    )
-    return data
-
-
-def _applications_sankey(response: Response) -> list[tuple[str, str, int]]:
-    """Provide a sankey data structure of applications, formatted for type checker."""
-    data = list(
-        chain(
-            *[
-                _data_product_application_sankey_links(data_product)
-                for data_product in response.data_products
-            ]
-        )
-    )
-    return data
-
-
-def _data_products_sankey(response: Response) -> list[tuple[str, str, int]]:
-    """Provide a sankey data structure of applications, formatted for type checker."""
-    data = list(
-        chain(
-            *[
-                _observing_system_data_product_sankey_links(observing_system)
-                for observing_system in response.observing_systems
-            ]
-        )
-    )
-    return data
-
-
-def _data_product_application_sankey_links(
-    data_product: ResponseDataProduct,
-) -> list[tuple[str, str, int]]:
-    data = [
-        (data_product.short_name, r.application.short_name, r.performance_rating)
-        for r in data_product.output_relationships
-    ]
-    return data
-
-
-def _observing_system_data_product_sankey_links(
-    observing_system: ResponseObservingSystem,
-) -> list[tuple[str, str, int]]:
-    data = [
-        (observing_system.short_name, r.data_product.short_name, r.performance_rating)
-        for r in observing_system.output_relationships
-    ]
-    return data
-
-
-def _application_societal_benefit_area_sankey_links(
-    application: ResponseApplication,
-) -> list[tuple[str, str, int]]:
-    data = [
-        (
-            application.short_name,
-            r.societal_benefit_area.societal_benefit_area_id,
-            r.performance_rating,
-        )
-        for r in application.output_relationships
-    ]
     return data
